globally:  code motion
