+++ cardano-node/src/Cardano/CLI/Genesis.hs
@@ -1,0 +1,71 @@
+{-# LANGUAGE CPP                        #-}
+{-# LANGUAGE RecordWildCards            #-}
+
+module Cardano.CLI.Genesis
+  ( dumpGenesis
+  )
+where
+
+import           Prelude (String)
+import           Cardano.Prelude hiding (option, show, trace)
+
+import qualified Data.ByteString.Lazy as LB
+import qualified Data.Map.Strict as Map
+import           Text.Printf (printf)
+
+import           System.Directory (createDirectory, doesPathExist)
+import           System.FilePath ((</>))
+#ifdef UNIX
+import           System.Posix.Files (ownerReadMode, setFileMode)
+#else
+import           System.Directory (emptyPermissions, readable, setPermissions)
+#endif
+
+import           Cardano.Chain.Delegation hiding (Map, epoch)
+import           Cardano.Chain.Genesis
+import           Cardano.Crypto (SigningKey (..))
+import qualified Cardano.Crypto.Signing as Crypto
+
+import           Cardano.CLI.Key
+import           Cardano.CLI.Ops
+
+
+dumpGenesis :: CLIOps IO -> FilePath -> GenesisData -> GeneratedSecrets -> IO ()
+dumpGenesis CLIOps{..} outDir genesisData GeneratedSecrets{..} = do
+  exists <- doesPathExist outDir
+  if exists
+    then throwIO $ OutputMustNotAlreadyExist outDir
+    else createDirectory outDir
+
+  let genesisJSONFile = outDir <> "/genesis.json"
+  LB.writeFile genesisJSONFile =<< coSerialiseGenesis genesisData
+
+  let dlgCertMap = unGenesisDelegation $ gdHeavyDelegation genesisData
+      isCertForSK :: SigningKey -> Certificate -> Bool
+      isCertForSK sk UnsafeACertificate{..} = delegateVK == Crypto.toVerification sk
+      findDelegateCert :: SigningKey -> IO Certificate
+      findDelegateCert sk =
+        case flip find (Map.elems dlgCertMap) . isCertForSK $ sk of
+          Nothing -> throwIO $ NoGenesisDelegationForKey $ prettySigningKeyPub sk
+          Just x  -> pure x
+      wOut :: String -> String -> (a -> IO LB.ByteString) -> [a] -> IO ()
+      wOut = writeSecrets outDir
+  dlgCerts <- mapM findDelegateCert gsRichSecrets
+
+  wOut "genesis-keys"    "key"  coSerialiseGenesisKey     gsDlgIssuersSecrets
+  wOut "delegate-keys"   "key"  coSerialiseDelegateKey    gsRichSecrets
+  wOut "poor-keys"       "key"  coSerialisePoorKey        gsPoorSecrets
+  wOut "delegation-cert" "json" coSerialiseDelegationCert dlgCerts
+  wOut "avvm-seed"       "seed" (pure . LB.fromStrict)     gsFakeAvvmSeeds
+
+writeSecrets :: FilePath -> String -> String -> (a -> IO LB.ByteString) -> [a] -> IO ()
+writeSecrets outDir prefix suffix secretOp xs =
+  forM_ (zip xs $ [0::Int ..]) $
+  \(secret, nr)-> do
+    let filename = outDir </> prefix <> "." <> printf "%03d" nr <> "." <> suffix
+    secretOp secret >>= LB.writeFile filename
+#ifdef UNIX
+    setFileMode                      filename ownerReadMode
+#else
+    setPermissions filename (emptyPermissions {readable = True})
+#endif
