+++ cardano-node/src/Cardano/CLI/Run.hs
@@ -1,5 +1,4 @@
 {-# LANGUAGE BangPatterns               #-}
-{-# LANGUAGE CPP                        #-}
 {-# LANGUAGE FlexibleContexts           #-}
 {-# LANGUAGE FlexibleInstances          #-}
 {-# LANGUAGE GADTs                      #-}
@@ -20,10 +19,6 @@
 {-# OPTIONS_GHC -Wno-all-missed-specialisations #-}
 {-# OPTIONS_GHC -Wno-simplifiable-class-constraints #-}
 
-#if !defined(mingw32_HOST_OS)
-#define UNIX
-#endif
-
 module Cardano.CLI.Run (
     CliError (..)
   , ClientCommand(..)
@@ -41,7 +36,7 @@ module Cardano.CLI.Run (
   , NewTxFile(..)
   ) where
 
-import           Prelude (String, error, show)
+import           Prelude (String, show)
 import           Cardano.Prelude hiding (option, show, trace)
 import           Test.Cardano.Prelude (canonicalDecodePretty)
 
@@ -49,41 +44,27 @@ import           Codec.Serialise (serialise, deserialiseOrFail)
 import           Control.Tracer
 import           Data.Bits (shiftL)
 import qualified Data.ByteArray as BA
-import qualified Data.ByteString as SB
 import qualified Data.ByteString.Lazy as LB
 import qualified Data.ByteString.UTF8 as UTF8
-import qualified Data.List.NonEmpty as NE
-import qualified Data.Map.Strict as Map
-import           Data.Map (Map)
 import           Data.Semigroup ((<>))
-import           Data.String (IsString, fromString)
+import           Data.String (IsString)
 import           Data.Text (Text)
 import qualified Data.Text as T
 import qualified Data.Text.Lazy as TL
 import qualified Data.Text.Lazy.IO as TL
 import qualified Data.Text.Lazy.Builder as Builder
-import qualified Data.Vector as V
 import qualified Formatting as F
-import           System.Directory (createDirectory, doesPathExist)
-import           System.FilePath ((</>))
+import           System.Directory (doesPathExist)
 import           System.IO (hGetLine, hSetEcho, hFlush, stdout, stdin)
-import           Text.Printf (printf)
 import           Data.Time (UTCTime)
-#ifdef UNIX
-import           System.Posix.Files (ownerReadMode, setFileMode)
-#else
-import           System.Directory (emptyPermissions, readable, setPermissions)
-#endif
 
-import           Cardano.Binary (Annotated(..), reAnnotate, serialize')
+import           Cardano.Binary (Annotated(..), serialize')
 import qualified Cardano.Chain.Common as CC.Common
 import           Cardano.Chain.Common
 import           Cardano.Chain.Delegation hiding (Map, epoch)
 import           Cardano.Chain.Genesis
-import qualified Cardano.Chain.Genesis as CC.Genesis
 import           Cardano.Chain.Slotting (EpochNumber(..))
 import           Cardano.Chain.UTxO
-import qualified Cardano.Chain.UTxO as CC.UTxO
 import           Cardano.Crypto (SigningKey (..), ProtocolMagic, ProtocolMagicId)
 import qualified Cardano.Crypto.Hashing as Crypto
 import qualified Cardano.Crypto.Random as Crypto
@@ -98,12 +79,15 @@ import           Ouroboros.Consensus.Ledger.Byron
 import           Ouroboros.Consensus.Ledger.Byron.Config
 import qualified Test.Cardano.Chain.Genesis.Dummy as Dummy
 
+import           Cardano.CLI.Genesis
+import           Cardano.CLI.Key
 import           Cardano.CLI.Ops
+import           Cardano.CLI.Tx
+import           Cardano.CLI.Tx.Submission
 import           Cardano.Config.CommonCLI
+import           Cardano.Common.Orphans ()
 import           Cardano.Common.Protocol
-import           Cardano.Node.Orphans ()
 import           Cardano.Node.Configuration.Topology
-import           Cardano.Node.TxSubmission
 
 
 newtype GenesisFile =
@@ -113,22 +97,6 @@ newtype GenesisFile =
 newtype NewDirectory =
   NewDirectory FilePath
   deriving (Eq, Ord, Show, IsString)
-
-newtype SigningKeyFile =
-  SigningKeyFile FilePath
-  deriving (Eq, Ord, Show, IsString)
-
-newtype NewSigningKeyFile =
-  NewSigningKeyFile FilePath
-  deriving (Eq, Ord, Show, IsString)
-
-newtype VerificationKeyFile =
-  VerificationKeyFile FilePath
-  deriving (Eq, Ord, Show, IsString)
-
-newtype NewVerificationKeyFile =
-  NewVerificationKeyFile FilePath
-   deriving (Eq, Ord, Show, IsString)
 
 newtype CertificateFile =
   CertificateFile FilePath
@@ -431,95 +399,6 @@ withRealPBFT CLIOps{coProtocol} pcc common action = do
       action cc proto
     _ -> throwIO $ ProtocolNotSupported coProtocol
 
-signTxId :: ProtocolMagicId -> SigningKey -> TxId -> CC.UTxO.TxInWitness
-signTxId pmid sk txid = CC.UTxO.VKWitness
-  (Crypto.toVerification sk)
-  (Crypto.sign
-    pmid
-    Crypto.SignTx
-    sk
-    (CC.UTxO.TxSigData txid))
-
-txSpendUTxOByronPBFT
-  :: CC.Genesis.Config
-  -> SigningKey
-  -> NonEmpty TxIn
-  -> NonEmpty TxOut
-  -> GenTx (ByronBlockOrEBB ByronConfig)
-txSpendUTxOByronPBFT gc sk ins outs =
-  let txattrs  = mkAttributes ()
-      tx       = UnsafeTx ins outs txattrs
-      wit      = signTxId (configProtocolMagicId gc) sk (Crypto.hash tx)
-      ATxAux atx awit =
-        mkTxAux tx . V.fromList . take (NE.length ins) $ repeat wit
-  in mkByronTx $ ATxAux (reAnnotate atx) (reAnnotate awit)
-
-txSpendGenesisUTxOByronPBFT
-  :: CC.Genesis.Config
-  -> SigningKey
-  -> Address
-  -> NonEmpty TxOut
-  -> GenTx (ByronBlockOrEBB ByronConfig)
-txSpendGenesisUTxOByronPBFT gc sk genAddr outs =
-  let txattrs  = mkAttributes ()
-      tx       = UnsafeTx (pure txIn) outs txattrs
-      txIn    :: CC.UTxO.TxIn
-      txIn     = genesisUTxOTxIn gc sk genAddr
-      wit      = signTxId (configProtocolMagicId gc) sk (Crypto.hash tx) 
-      ATxAux atx awit = mkTxAux tx . V.fromList . pure $ wit
-  in mkByronTx $ ATxAux (reAnnotate atx) (reAnnotate awit)
-
--- | Given a genesis, and a pair of signing key and address, reconstruct a TxIn
---   corresponding to the genesis UTxO entry.
-genesisUTxOTxIn :: CC.Genesis.Config -> SigningKey -> Address -> CC.UTxO.TxIn
-genesisUTxOTxIn gc genSk genAddr =
-  let vk         = Crypto.toVerification genSk
-      handleMissingAddr :: Maybe CC.UTxO.TxIn -> CC.UTxO.TxIn
-      handleMissingAddr  = fromMaybe . error
-        $  "\nGenesis UTxO has no address\n"
-        <> (T.unpack $ prettyAddress genAddr)
-        <> "\n\nIt has the following, though:\n\n"
-        <> Cardano.Prelude.concat (T.unpack . prettyAddress <$> Map.keys initialUtxo)
-
-      initialUtxo :: Map Address (CC.UTxO.TxIn, CC.UTxO.TxOut)
-      initialUtxo =
-            Map.fromList
-          . mapMaybe (\(inp, out) -> mkEntry inp genAddr <$> keyMatchesUTxO vk out)
-          . fromCompactTxInTxOutList
-          . Map.toList
-          . CC.UTxO.unUTxO
-          . CC.UTxO.genesisUtxo
-          $ gc
-        where
-          mkEntry :: CC.UTxO.TxIn
-                  -> Address
-                  -> CC.UTxO.TxOut
-                  -> (Address, (CC.UTxO.TxIn, CC.UTxO.TxOut))
-          mkEntry inp addr out = (addr, (inp, out))
-
-      keyMatchesUTxO :: Crypto.VerificationKey -> CC.UTxO.TxOut -> Maybe CC.UTxO.TxOut
-      keyMatchesUTxO key out =
-        if CC.Common.checkVerKeyAddress key (CC.UTxO.txOutAddress out)
-        then Just out else Nothing
-
-      fromCompactTxInTxOutList :: [(CC.UTxO.CompactTxIn, CC.UTxO.CompactTxOut)]
-                               -> [(CC.UTxO.TxIn, CC.UTxO.TxOut)]
-      fromCompactTxInTxOutList =
-          map (bimap CC.UTxO.fromCompactTxIn CC.UTxO.fromCompactTxOut)
-  in handleMissingAddr $ fst <$> Map.lookup genAddr initialUtxo
-
--- TODO:  we need to support password-protected secrets.
-readSigningKey :: CLIOps IO -> SigningKeyFile -> IO SigningKey
-readSigningKey co (SigningKeyFile fp) =
-  coDeserialiseDelegateKey co fp =<< LB.readFile fp
-
-readVerificationKey :: VerificationKeyFile -> IO Crypto.VerificationKey
-readVerificationKey (VerificationKeyFile fp) = do
-  vkB <- SB.readFile fp
-  case Crypto.parseFullVerificationKey . fromString $ UTF8.toString vkB of
-    Left e -> throwIO . VerificationKeyDeserialisationFailed fp $ T.pack $ show e
-    Right x -> pure x
-
 -- TODO:  we'd be better served by a combination of a temporary file
 --        with an atomic rename.
 
@@ -554,53 +433,3 @@ readPassword prompt = do
           else hPutStrLn stdout ("Sorry, entered passwords don't match." :: String)
                >> loop
   Crypto.PassPhrase . BA.convert . UTF8.fromString <$> loop
-
-dumpGenesis :: CLIOps IO -> FilePath -> GenesisData -> GeneratedSecrets -> IO ()
-dumpGenesis CLIOps{..} outDir genesisData GeneratedSecrets{..} = do
-  exists <- doesPathExist outDir
-  if exists
-    then throwIO $ OutputMustNotAlreadyExist outDir
-    else createDirectory outDir
-
-  let genesisJSONFile = outDir <> "/genesis.json"
-  LB.writeFile genesisJSONFile =<< coSerialiseGenesis genesisData
-
-  let dlgCertMap = unGenesisDelegation $ gdHeavyDelegation genesisData
-      isCertForSK :: SigningKey -> Certificate -> Bool
-      isCertForSK sk UnsafeACertificate{..} = delegateVK == Crypto.toVerification sk
-      findDelegateCert :: SigningKey -> IO Certificate
-      findDelegateCert sk =
-        case flip find (Map.elems dlgCertMap) . isCertForSK $ sk of
-          Nothing -> throwIO $ NoGenesisDelegationForKey $ prettySigningKeyPub sk
-          Just x  -> pure x
-      wOut :: String -> String -> (a -> IO LB.ByteString) -> [a] -> IO ()
-      wOut = writeSecrets outDir
-  dlgCerts <- mapM findDelegateCert gsRichSecrets
-
-  wOut "genesis-keys"    "key"  coSerialiseGenesisKey     gsDlgIssuersSecrets
-  wOut "delegate-keys"   "key"  coSerialiseDelegateKey    gsRichSecrets
-  wOut "poor-keys"       "key"  coSerialisePoorKey        gsPoorSecrets
-  wOut "delegation-cert" "json" coSerialiseDelegationCert dlgCerts
-  wOut "avvm-seed"       "seed" (pure . LB.fromStrict)     gsFakeAvvmSeeds
-
-prettySigningKeyPub :: SigningKey -> Text
-prettySigningKeyPub (Crypto.toVerification -> vk) = TL.toStrict
-  $  "public key hash: " <> (F.format Crypto.hashHexF . CC.Common.addressHash $ vk) <> "\n"
-  <> "     public key: " <> (Builder.toLazyText . Crypto.formatFullVerificationKey $ vk)
-
-prettyAddress :: CC.Common.Address -> Text
-prettyAddress addr = TL.toStrict
-  $  F.format CC.Common.addressF         addr <> "\n"
-  <> F.format CC.Common.addressDetailedF addr
-
-writeSecrets :: FilePath -> String -> String -> (a -> IO LB.ByteString) -> [a] -> IO ()
-writeSecrets outDir prefix suffix secretOp xs =
-  forM_ (zip xs $ [0::Int ..]) $
-  \(secret, nr)-> do
-    let filename = outDir </> prefix <> "." <> printf "%03d" nr <> "." <> suffix
-    secretOp secret >>= LB.writeFile filename
-#ifdef UNIX
-    setFileMode                      filename ownerReadMode
-#else
-    setPermissions filename (emptyPermissions {readable = True})
-#endif
