+++ cardano-node/src/Cardano/Common/LocalSocket.hs
@@ -1,0 +1,25 @@
+module Cardano.Common.LocalSocket
+  ( localSocketFilePath
+  , localSocketAddrInfo
+  )
+where
+
+import           Cardano.Prelude
+
+import           Network.Socket as Socket
+import           Ouroboros.Consensus.NodeId
+
+
+localSocketFilePath :: NodeId -> FilePath
+localSocketFilePath (CoreId  n) = "node-core-" ++ show n ++ ".socket"
+localSocketFilePath (RelayId n) = "node-relay-" ++ show n ++ ".socket"
+
+localSocketAddrInfo :: FilePath -> Socket.AddrInfo
+localSocketAddrInfo socketPath =
+    Socket.AddrInfo
+      []
+      Socket.AF_UNIX
+      Socket.Stream
+      Socket.defaultProtocol
+      (Socket.SockAddrUnix socketPath)
+      Nothing
